feat(frontend): review adjustments
