+++ src/frontend/src/btc/components/send/BtcSendReview.svelte
@@ -9,6 +9,7 @@
 	} from '$btc/derived/btc-pending-sent-transactions-status.derived';
 	import type { UtxosFee } from '$btc/types/btc-send';
 	import SendReview from '$lib/components/send/SendReview.svelte';
+	import { ProgressStepsSendBtc } from '$lib/enums/progress-steps';
 	import type { NetworkId } from '$lib/types/network';
 	import { invalidAmount } from '$lib/utils/input.utils';
 	import { isInvalidDestinationBtc } from '$lib/utils/send.utils';
@@ -18,12 +19,13 @@
 	export let networkId: NetworkId | undefined = undefined;
 	export let source: string;
 	export let utxosFee: UtxosFee | undefined = undefined;
+	export let progress: (step: ProgressStepsSendBtc) => void;
 
 	let hasPendingTransactionsStore: Readable<BtcPendingSentTransactionsStatus>;
 	$: hasPendingTransactionsStore = initPendingSentTransactionsStatus(source);
 
 	let disableSend: boolean;
-	// We want to disable send if pending transactions or UTXOs fee are loading, there was an error or there are pending transactions.
+	// We want to disable send if pending transactions or UTXOs fee isn't available yet, there was an error or there are pending transactions.
 	$: disableSend =
 		$hasPendingTransactionsStore !== BtcPendingSentTransactionsStatus.NONE ||
 		isNullish(utxosFee) ||
@@ -39,7 +41,7 @@
 </script>
 
 <SendReview on:icBack on:icSend {source} {amount} {destination} disabled={disableSend}>
-	<BtcUtxosFee slot="fee" on:icSelectUtxosFee {utxosFee} />
+	<BtcUtxosFee slot="fee" bind:utxosFee {progress} {networkId} {amount} />
 
 	<BtcSendHasPendingTransactions
 		slot="info"
